# feat: vector e4 utils ops