--- a/field/babybear/extensions/vector.go
+++ b/field/babybear/extensions/vector.go
@@ -248,6 +248,7 @@ func (vector *Vector) WriteTo(w io.Writer) (int64, error) {
 	if err != nil {
 		return n, err
 	}
+
 	return n, nil
 }