--- a/field/generator/generator_extensions.go
+++ b/field/generator/generator_extensions.go
@@ -57,6 +57,7 @@ func generateExtensions(F *config.Field, outputDir string) error {
 	if F.F31 {
 		entries_ext4 := []bavard.Entry{
 			{File: filepath.Join(outputDir, "e4.go"), Templates: []string{"e4.go.tmpl"}},
+			{File: filepath.Join(outputDir, "vector.go"), Templates: []string{"vector.go.tmpl"}},
 			{File: filepath.Join(outputDir, "e4_test.go"), Templates: []string{"e4_test.go.tmpl"}},
 		}