… field/babybear/extensions/e4.go
… field/generator/internal/templates/extensions/e4.go.tmpl
… field/generator/internal/templates/extensions/e4_test.go.tmpl
… field/koalabear/extensions/e4.go
… field/koalabear/extensions/e4_test.go