# Description

Bring back e4 vector serialization ops from linea monorepo and improve them:

```
benchmark                                                old ns/op     new ns/op     delta
BenchmarkVectorSerialization/MarshalBinary-96            1276132       574954        -54.95%
BenchmarkVectorSerialization/MarshalBinary-96            1192971       566853        -52.48%
BenchmarkVectorSerialization/UnmarshalBinary-96          1184194       333680        -71.82%
BenchmarkVectorSerialization/UnmarshalBinary-96          1183500       335667        -71.64%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     669630        329621        -50.78%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     673024        328453        -51.20%

benchmark                                                old allocs     new allocs     delta
BenchmarkVectorSerialization/MarshalBinary-96            18             5              -72.22%
BenchmarkVectorSerialization/MarshalBinary-96            18             5              -72.22%
BenchmarkVectorSerialization/UnmarshalBinary-96          3              6              +100.00%
BenchmarkVectorSerialization/UnmarshalBinary-96          3              6              +100.00%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     7              6              -14.29%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     7              6              -14.29%

benchmark                                                old bytes     new bytes     delta
BenchmarkVectorSerialization/MarshalBinary-96            2097161       1056894       -49.60%
BenchmarkVectorSerialization/MarshalBinary-96            2097162       1056895       -49.60%
BenchmarkVectorSerialization/UnmarshalBinary-96          524346        524547        +0.04%
BenchmarkVectorSerialization/UnmarshalBinary-96          524346        524541        +0.04%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     524579        524537        -0.01%
BenchmarkVectorSerialization/unmarshalBinaryAsync-96     524574        524534        -0.01%
```